requirements.txt
asgiref==3.8.1
Django==5.0.7
Faker==12.0.1
fpdf==1.7.2
lxml==5.3.0
mixer==7.2.2
numpy==2.0.1
pdf2image==1.17.0
pillow==10.4.0
pygame==2.6.0
PyPDF2==3.0.1
pyserial==3.5
python-dateutil==2.9.0.post0
python-docx==1.1.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1